Fix compilation with dotty by splitting classes depending on Global
Some classes cannot be compiled as part of dotty since we don't compile
Global, but we still need to compiler their companion objects because
they are referenced in other parts of the backend. We address this by
moving these objects in separate files that can be compiled by both
scalac and dotty, whereas the part that depend on Global will not be
compiled by dotty, this is achieved by using filters in sbt:

unmanagedSourceDirectories in Compile ++= {
  val backendDir = baseDirectory.value / ".." / "scala-backend" / "src" / "compiler" / "scala" / "tools" / "nsc" / "backend"
  val allScalaFiles = GlobFilter("*.scala")
  ((backendDir *
    (allScalaFiles - "JavaPlatform.scala" - "Platform.scala" - "ScalaPrimitives.scala")) +++
   (backendDir / "jvm") *
    (allScalaFiles - "BCodeICodeCommon.scala" - "GenASM.scala" - "GenBCode.scala" - "ScalacBackendInterface.scala")
  ).get
},

Summary of the changes:
- ScalaPrimitives is split into ScalaPrimitivesOps (shared) and
  ScalaPrimitives (scalac-only)
- jvm.GenBCode is split into jvm.GenBCodeOps (shared) and
  jvm.GenBCode (scalac-only)
- icode.Opcodes is split into jvm.Opcodes (shared) and
  icode.Opcodes (scalac-only)
- icode.Primitives is split into jvm.Primitives (shared) and
  icode.Primitives (scalac-only)
