+++ src/compiler/scala/tools/nsc/backend/jvm/BCodeAsmCommon.scala
@@ -6,7 +6,6 @@
 package scala.tools.nsc.backend.jvm
 
 import scala.reflect.internal.AnnotationInfos
-import scala.tools.nsc.Global
 
 /**
  * This trait contains code shared between GenBCode and GenASM that depends on types defined in
