+++ src/compiler/scala/tools/nsc/backend/jvm/BTypes.scala
@@ -7,7 +7,6 @@ package scala.tools.nsc
 package backend.jvm
 
 import scala.tools.asm
-import asm.Opcodes
 
 /**
  * The BTypes component defines The BType class hierarchy. BTypes encapsulates all type information
@@ -216,7 +215,7 @@ abstract class BTypes {
      *         `opcode` is `IRETURN`, this method returns `FRETURN`.
      */
     final def typedOpcode(opcode: Int): Int = {
-      if (opcode == Opcodes.IALOAD || opcode == Opcodes.IASTORE)
+      if (opcode == asm.Opcodes.IALOAD || opcode == asm.Opcodes.IASTORE)
         opcode + loadStoreOpcodeOffset
       else
         opcode + typedOpcodeOffset
@@ -664,7 +663,7 @@ abstract class BTypes {
           internalName,
           outerName.orNull,
           innerName.orNull,
-          GenBCode.mkFlags(
+          GenBCodeOps.mkFlags(
             info.flags,
             if (isStaticNestedClass) asm.Opcodes.ACC_STATIC else 0
           ) & ClassBType.INNER_CLASSES_FLAGS
