+++ src/compiler/scala/tools/nsc/backend/jvm/CoreBTypes.scala
@@ -164,7 +164,7 @@ class CoreBTypes[BTFS <: BTypesFromSymbols[_ <: BackendInterface]](val bTypes: B
   )
 
   lazy val typeOfArrayOp: Map[Int, BType] = {
-    import scala.tools.nsc.backend.ScalaPrimitives._
+    import scala.tools.nsc.backend.ScalaPrimitivesOps._
     Map(
         (List(ZARRAY_LENGTH, ZARRAY_GET, ZARRAY_SET) map (_ -> BOOL))   ++
         (List(BARRAY_LENGTH, BARRAY_GET, BARRAY_SET) map (_ -> BYTE))   ++
