+++ src/compiler/scala/tools/nsc/backend/jvm/GenASM.scala
@@ -9,13 +9,13 @@ package backend.jvm
 
 import scala.collection.{ mutable, immutable }
 import scala.reflect.internal.pickling.{ PickleFormat, PickleBuffer }
-import scala.tools.nsc.backend.icode.{TypeKinds, Primitives}
+import scala.tools.nsc.backend.icode.TypeKinds
 import scala.tools.nsc.symtab._
 import scala.tools.asm
 import asm.Label
 import scala.annotation.tailrec
 import BCodeAsmCommon._
-import scala.tools.nsc.backend.icode.Opcodes._
+import Opcodes._
 import Primitives._
 
 /**
