+++ src/compiler/scala/tools/nsc/backend/jvm/GenBCode.scala
@@ -410,9 +410,5 @@ abstract class GenBCode extends SubComponent with BCodeSyncAndTry {
 
 } // end of class GenBCode
 
-object GenBCode {
-  def mkFlags(args: Int*) = args.foldLeft(0)(_ | _)
 
-  final val PublicStatic      = asm.Opcodes.ACC_PUBLIC | asm.Opcodes.ACC_STATIC
-  final val PublicStaticFinal = asm.Opcodes.ACC_PUBLIC | asm.Opcodes.ACC_STATIC | asm.Opcodes.ACC_FINAL
-}
+object GenBCode extends GenBCodeOps
