+++ src/compiler/scala/tools/nsc/backend/opt/DeadCodeElimination.scala
@@ -16,8 +16,8 @@ abstract class DeadCodeElimination extends SubComponent {
   import icodes._
   import icodes.opcodes._
   import definitions.RuntimePackage
-  import scala.tools.nsc.backend.icode.Primitives._
-  import scala.tools.nsc.backend.icode.Opcodes._
+  import scala.tools.nsc.backend.jvm.Primitives._
+  import scala.tools.nsc.backend.jvm.Opcodes._
 
   /** The block and index where an instruction is located */
   type InstrLoc = (BasicBlock, Int)
