+++ src/compiler/scala/tools/nsc/backend/opt/Inliners.scala
@@ -41,8 +41,8 @@ abstract class Inliners extends SubComponent {
   import global._
   import icodes._
   import icodes.opcodes._
-  import scala.tools.nsc.backend.icode.Primitives._
-  import scala.tools.nsc.backend.icode.Opcodes._
+  import scala.tools.nsc.backend.jvm.Primitives._
+  import scala.tools.nsc.backend.jvm.Opcodes._
 
   import definitions.{
     NullClass, NothingClass, ObjectClass,
